chore: clear the clippy backlog and gate CI on it
CI ran `cargo check` and `cargo test` but never clippy, so roughly 150
warnings had accumulated across the workspace. This clears them and adds
`cargo clippy --workspace --all-targets -- -D warnings` to CI so the count
cannot drift back up.

Most were mechanical (`map_or` -> `is_some_and`, redundant closures,
needless range loops, manual prefix stripping, `x >= y + 1` -> `x > y`).
Three needed judgement:

- Six `if` statements had two identical branches whose comments explained
  a strand distinction that the code did not actually make. The branches
  are collapsed and the explanation kept, since the value genuinely does
  not depend on strand there.
- Two doc comments were rendering as blockquotes because a line began
  with `>` continuing a code span (`>= 2 BP`, `> 0`). Reflowed. One of
  them also documented BS2's old recessive rule, which no longer holds,
  so it is rewritten to point at the null-individual test.
- The repeated `(Option<(String, String)>, Option<(String, String)>)` in
  the consequence predictor is now a named `DisplayPair`.

Where a lint is wrong for a site, it is silenced there with an `#[allow]`
and a one-line reason rather than globally: functions whose arguments are
independent coordinates with no natural grouping, and inherent
constructors (`from_str`, `from_iter`) that deliberately do not match the
fallible or generic std trait they resemble.

749 tests pass; no behaviour change intended.

+++ crates/fastvep-sa/tests/round_trip.rs
@@ -227,7 +227,7 @@ fn test_osa2_writer_reader_round_trip() {
                 alt_allele: b"G".to_vec(),
                 values: vec![
                     fields[0].encode_float(af), // AF
-                    (i + 1) as u32,             // AC
+                    (i + 1),               // AC
                 ],
                 json_blob: None,
             }
@@ -314,7 +314,7 @@ fn test_osa2_preload() {
             position: 1000 + i * 10,
             ref_allele: b"C".to_vec(),
             alt_allele: b"T".to_vec(),
-            values: vec![i as u32 + 1],
+            values: vec![i + 1],
             json_blob: None,
         })
         .collect();
